Refactor processPatents.js into modular structure
This commit refactors the monolithic processPatents.js into multiple modules:
- PatentParser: Handles parsing of patent data from text
- PatentCategorizer: Handles categorization of patents
- SubcategoryProcessor: Handles subcategory classification
- FileHandler: Manages file operations
- categoryConfig.js: Contains category definitions and keywords

The original processPatents.js now acts as a wrapper
that imports and exports from the modular implementation,
ensuring backward compatibility with existing code.
